+++ examples/with-first-chunk-cache-ssr/src/app.tsx
@@ -1,0 +1,10 @@
+import { defineAppConfig } from 'ice';
+
+export default defineAppConfig({
+  app: {
+    rootId: 'app',
+  },
+  cache: {
+    firstChunk: true,
+  },
+});